requirements.txt
rich>=13.7,<14
click>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
dnspython>=2.0.0
openai>=1.0.0
requests>=2.25.0
cryptography>=3.4.0
pycryptodome>=3.10.0
colorama>=0.4.0
psutil>=5.8.0